Added Tasks view

+++ src/main/java/ca/jusjoken/data/service/TaskPlanRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import ca.jusjoken.data.entity.TaskPlan;
 
 @Repository
-public interface TaskPlanRepository extends JpaRepository<TaskPlan, Long> {
+public interface TaskPlanRepository extends JpaRepository<TaskPlan, Integer> {
 }
